viui::event

Enum Event

pub enum Event {
    MouseMove {
        pos: Point,
    },
    MousePress {
        pos: Point,
        button: MouseButton,
    },
    MouseRelease {
        pos: Point,
        button: MouseButton,
    },
    Scroll {
        pos: Point,
        delta_y: f32,
    },
    TouchBegin {
        pos: Point,
        finger_id: u32,
    },
    TouchMove {
        pos: Point,
        finger_id: u32,
    },
    TouchEnd {
        pos: Point,
        finger_id: u32,
    },
    KeyPress {
        key: KeyCode,
        modifiers: Modifiers,
    },
    KeyRelease {
        key: KeyCode,
    },
    Char(char),
    Focus,
    Blur,
}

All input events delivered to widgets.

Variants

MouseMove

Fields

pos: Point

MousePress

Fields

pos: Point

button: MouseButton

MouseRelease

Also dispatched globally (GlobalRelease strategy) to clear PRESSED state.

Fields

pos: Point

button: MouseButton

Scroll

Fields

pos: Point

delta_y: f32

TouchBegin

First contact of a new touch (maps to MousePress for non-touch widgets).

Fields

pos: Point

finger_id: u32

TouchMove

Contact moved while touching (maps to MouseMove for non-touch widgets).

Fields

pos: Point

finger_id: u32

TouchEnd

Contact lifted (maps to MouseRelease for non-touch widgets).

Fields

pos: Point

finger_id: u32

KeyPress

Fields

key: KeyCode

modifiers: Modifiers

KeyRelease

Fields

key: KeyCode

Char(char)

Focus

Blur

Implementations

impl Event

pub fn pointer_pos(&self) -> Option<Point>

Returns the pointer position for pointer events, None for keyboard/lifecycle.

pub fn is_global_release(&self) -> bool

True if this event should be dispatched globally (not just to the hit-target).